requirements.txt
pillow
django
djangorestframework
rq
wand